Moving the TypeScript solution to separate folder.
